Move 'src/domain/mutations' to 'src/common/mutations'
